ColorSchemeProvider, OnLinkNavigationProvider, useFocusVisible, useReducedMotion: convert examples to Sandpack
